changed to calculate with a float to give 0.5 in case render times is 1

+++ libgdx-test-apps-core/src/main/java/com/gemserk/libgdx/tests/BlendingPerformanceTestScreen.java
@@ -316,7 +316,7 @@ public class BlendingPerformanceTestScreen extends TestScreen {
 		else
 			spriteBatch.disableBlending();
 
-		int rt = !twoTextures ? renderTimes : Math.round(renderTimes / 2);
+		int rt = !twoTextures ? renderTimes : Math.round(((float)renderTimes) / 2f);
 		
 		if (!multipleBatches) {
 			spriteBatch.begin();
